Replace fprintf with fprintf_alternative
fprintf has been included in the final binaries after assertions has
been enabled (but I can't be 100% certain - this must be verified, though,
with enabling/disabling assertions in libc++ and libc++abi).

To reduce the library size I've changed the name of fprintf and vfprintf
to fprintf_alternative and vfprintf_alternative, correspondingly. The
functions are defined weak, so the user can change them.