Fix linking to operator new(std::size_t, std::align_val_t)
C++17 introduced a new 'new' operator:

    void* operator new  ( std::size_t count, std::align_val_t al );

We don't have it in ARM baremetal, because there is no posix_memalign()
implemented for that architecture. This operator 'new' function needs
posix_memalign().

Compiling the libraries with C++14 disables the calls to the new 'new'
operator overload, thus the dependants doesn't have to provide a custom
implementation of the above overload.