Install newlib libc
See the README.md for explanation.
Regarding the implementation: we use arm-none-eabi-gcc, bundled with ARM
GNU Toolchain, to figure out what is the 'libc.a' path, for the
specified architecture flags. We parse it to retrieve the path not only
to libc.a, but also to other libs: libg.a, rdimon, ...