Include subarchitecture and float handling info in package name
The package name shall include all the relevant info, to find out what
does it contain. Thus, we include the subarchitecture name (cortex-m*)
and the floating point unit, if applicable, or 'soft_float' if not, in
the package name. We retrieve it from the target compile flags, from the
LLVM_BAREMETAL_ARM_TARGET_COMPILE_FLAGS  variable. The info, whether
exceptions are enabled for that build is also included.